requirements.txt
requests==2.31.0
requests_oauthlib==1.3.1
python-dotenv==1.0.0
cryptography==44.0.1
brother-ql==0.9.4